requirements.txt
fastapi==0.115.12
numpy==2.2.5
onnxruntime==1.21.1
opencv_python==4.10.0.84
Pillow==11.2.1
pydantic==2.11.4
PyYAML==6.0.1
PyYAML==6.0.2
